app_ml_otodom/requirements.txt
pip
lxml
boto3
requests
pymongo==3.9.0
Pillow
pandas==0.25.1
botocore
numpy==1.15.4
xmltodict
Scrapy==1.7.3
twisted
geopandas
pymongo
kafka-python
scikit-learn==0.20.0
flask==0.12
